refactoring + fix sessions
